[#23] Make RollbarThread more correct
RollbarThread was doing a few things that could lead to deadlocks by
misusing Java concurrency primatives. Testing a deadlock scenario is
pretty hard usually, especially because I am not exactly sure which of
the oddities was necessarily the culprit. Nonetheless, this is more
idiomatic Java threading/locking code.

Looking at the stack trace in the issue, it looks like the handling of
interrupts is the probable cause.